// src/shared/api/generated/models/disableSupportCaseNotificationPolicyDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export interface DisableSupportCaseNotificationPolicyDto {
  /**
   * @minimum 0
   * @maximum 2147483647
   */
  expectedVersion: number;
  /**
   * @minLength 3
   * @maxLength 500
   */
  reason: string;
}